# internal/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nops API",
    "version": "1",
    "license": { "name": "Apache 2.0", "identifier": "Apache-2.0" },
    "description": "Read jobs and deployments, act on them, and manage ACL policies, binding rules and tokens. A token does what its ACL policies allow, with the dashboard's rules."
  },
  "servers": [{ "url": ".." }],
  "security": [{ "bearer": [] }],
  "paths": {
...
    "/api/jobs": {
      "get": {
        "summary": "Every managed job, orphans included",
        "operationId": "listJobs",
        "responses": {
          "200": {
            "description": "The jobs, by namespace and name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["jobs"],
                  "properties": { "jobs": { "type": "array", "items": { "$ref": "#/components/schemas/Job" } } }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/jobs/{namespace}/{job}": {
      "get": {
        "summary": "One job with its drift, hold, and deployments",
        "operationId": "getJob",
        "parameters": [{ "$ref": "#/components/parameters/Namespace" }, { "$ref": "#/components/parameters/Job" }],
        "responses": {
          "200": {
            "description": "The job.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobDetail" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/deployments/{id}": {
      "get": {
        "summary": "One deployment with its redacted diff, hook runs, and events",
        "operationId": "getDeployment",
        "parameters": [{ "$ref": "#/components/parameters/DeploymentID" }],
        "responses": {
          "200": {
            "description": "The deployment.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeploymentDetail" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/deployments/{id}/approve": {
      "post": {
        "summary": "Approve a deployment",
        "operationId": "approveDeployment",
        "parameters": [{ "$ref": "#/components/parameters/DeploymentID" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SpecHash" } } }
        },
        "responses": {
          "204": { "description": "Approved." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/deployments/{id}/reject": {
      "post": {
        "summary": "Reject a deployment",
        "operationId": "rejectDeployment",
        "parameters": [{ "$ref": "#/components/parameters/DeploymentID" }],
        "responses": {
          "204": { "description": "Rejected." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/deployments/{id}/promote": {
      "post": {
        "summary": "Promote the canaries of a deployment that waits for it",
        "operationId": "promoteDeployment",
        "parameters": [{ "$ref": "#/components/parameters/DeploymentID" }],
        "responses": {
          "204": { "description": "Promoted." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/deployments/{id}/retry": {
      "post": {
        "summary": "Retry a failed or rejected deployment",
        "operationId": "retryDeployment",
        "parameters": [{ "$ref": "#/components/parameters/DeploymentID" }],
        "responses": {
          "200": {
            "description": "The new deployment.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeploymentID" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/jobs/{namespace}/{job}/pause": {
      "post": {
        "summary": "Pause a job",
        "operationId": "pauseJob",
        "parameters": [{ "$ref": "#/components/parameters/Namespace" }, { "$ref": "#/components/parameters/Job" }],
        "requestBody": {
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Pause" } } }
        },
        "responses": {
          "204": { "description": "Paused." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/jobs/{namespace}/{job}/resume": {
      "post": {
        "summary": "Resume a paused job",
        "operationId": "resumeJob",
        "parameters": [{ "$ref": "#/components/parameters/Namespace" }, { "$ref": "#/components/parameters/Job" }],
        "responses": {
          "204": { "description": "Resumed." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/jobs/{namespace}/{job}/deploy-now": {
      "post": {
        "summary": "Deploy a job outside its sync window",
        "operationId": "deployJobNow",
        "parameters": [{ "$ref": "#/components/parameters/Namespace" }, { "$ref": "#/components/parameters/Job" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SpecHash" } } }
        },
        "responses": {
          "200": {
            "description": "The deployment the request led to. Its id is empty when the next detection cycle starts it.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeploymentID" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": { "$ref": "#/components/responses/Conflict" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/fetch": {
      "post": {
        "summary": "Ask for a git poll now",
        "operationId": "fetchGit",
        "responses": {
          "202": { "description": "The poll is asked for; it runs on its own." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/api/acl/policies/{name}": {
      "get": {
        "summary": "One ACL policy with its rules",
        "operationId": "getACLPolicy",
        "parameters": [{ "$ref": "#/components/parameters/PolicyName" }],
        "responses": {
          "200": {
            "description": "The ACL policy.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ACLPolicy" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "put": {
        "summary": "Create an ACL policy or replace it",
        "operationId": "putACLPolicy",
        "parameters": [{ "$ref": "#/components/parameters/PolicyName" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ACLPolicyBody" } } }
        },
        "responses": {
          "200": {
            "description": "The ACL policy as saved.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ACLPolicy" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "summary": "Delete an ACL policy",
        "operationId": "deleteACLPolicy",
        "parameters": [{ "$ref": "#/components/parameters/PolicyName" }],
        "responses": {
          "204": { "description": "Deleted." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/acl/binding-rules/{id}": {
      "get": {
        "summary": "One binding rule",
        "operationId": "getBindingRule",
        "parameters": [{ "$ref": "#/components/parameters/BindingRuleID" }],
        "responses": {
          "200": {
            "description": "The binding rule.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BindingRule" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "put": {
        "summary": "Replace a binding rule",
        "operationId": "putBindingRule",
        "parameters": [{ "$ref": "#/components/parameters/BindingRuleID" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BindingRuleBody" } } }
        },
        "responses": {
          "200": {
            "description": "The binding rule as saved.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BindingRule" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "summary": "Delete a binding rule",
        "operationId": "deleteBindingRule",
        "parameters": [{ "$ref": "#/components/parameters/BindingRuleID" }],
        "responses": {
          "204": { "description": "Deleted." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
    "/api/acl/tokens": {
      "get": {
        "summary": "Every token",
        "operationId": "listACLTokens",
        "responses": {
          "200": {
            "description": "The tokens, newest first.",
            "content": { "application/json": { "schema": { "type": "object", "required": ["tokens"], "properties": { "tokens": { "type": "array", "items": { "$ref": "#/components/schemas/ACLToken" } } } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "post": {
        "summary": "Create a token",
        "operationId": "createACLToken",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ACLTokenBody" } } }
        },
        "responses": {
          "201": {
            "description": "The token, with its secret.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/NewACLToken" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
    "/api/acl/tokens/{accessor_id}": {
      "get": {
        "summary": "One token",
        "operationId": "getACLToken",
        "parameters": [{ "$ref": "#/components/parameters/AccessorID" }],
        "responses": {
          "200": {
            "description": "The token.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ACLToken" } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      },
      "delete": {
        "summary": "Revoke a token",
        "operationId": "revokeACLToken",
        "parameters": [{ "$ref": "#/components/parameters/AccessorID" }],
        "responses": {
          "204": { "description": "Revoked." },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/ServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": { "type": "http", "scheme": "bearer", "description": "A token from the dashboard's Administration page." }
    },
    "parameters": {
      "Namespace": { "name": "namespace", "in": "path", "required": true, "schema": { "type": "string" } },
      "Job": { "name": "job", "in": "path", "required": true, "schema": { "type": "string" } },
      "DeploymentID": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
...
    },
    "responses": {
      "BadRequest": {
        "description": "The request is not what the endpoint takes.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "No valid token: it is missing, unknown, expired, or revoked.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Forbidden": {
        "description": "The token is valid, and its ACL policies do not allow this.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "The job, deployment, ACL policy or token does not exist.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Conflict": {
        "description": "The state moved since you read it.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ServerError": {
        "description": "Nops hit an error and logged it.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
...
      "SpecHash": {
        "type": "object",
        "required": ["spec_hash"],
        "properties": { "spec_hash": { "type": "string", "description": "The spec hash you read." } }
      },
      "Pause": {
        "type": "object",
        "properties": { "reason": { "type": "string", "maxLength": 500, "description": "At most 500 bytes." } }
      },
      "DeploymentID": {
        "type": "object",
        "required": ["deployment_id"],
        "properties": { "deployment_id": { "type": "string" } }
      },
...
      "Job": {
        "type": "object",
        "required": ["namespace", "job", "policy", "sync", "drift"],
        "properties": {
          "namespace": { "type": "string" },
          "job": { "type": "string" },
          "policy": { "type": "string", "enum": ["auto", "approval", "none"] },
          "sync": {
            "type": "string",
            "description": "The sync state of the job: invalid, paused, blocked, orphan, pending, deploying, held, drift, or sync. Empty for a job gone from git and from Nomad."
          },
          "drift": { "type": "boolean" },
          "spec_hash": { "type": "string", "description": "What a deployment of the job would have now: the hash Deploy now asks for." },
          "blocked_by": { "type": "string" },
          "hold": { "$ref": "#/components/schemas/Hold" },
          "last_deployment": { "$ref": "#/components/schemas/Deployment" }
        }
      },
...
      "JobDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Job" },
          {
            "type": "object",
            "required": ["deployments"],
            "properties": {
              "file": { "type": "string" },
              "diff": { "type": "object", "description": "The redacted plan diff of the drift, as Nomad's JSON. Absent without drift." },
              "issues": { "type": "array", "items": { "$ref": "#/components/schemas/Issue" } },
              "deployments": { "type": "array", "items": { "$ref": "#/components/schemas/Deployment" } }
            }
          }
        ]
      },
...
      "DeploymentDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/Deployment" },
          {
            "type": "object",
            "required": ["events", "hook_runs"],
            "properties": {
              "diff": { "type": "object", "description": "The redacted plan diff, as Nomad's JSON. Absent when there is none." },
              "events": { "type": "array", "items": { "$ref": "#/components/schemas/Event" } },
              "hook_runs": { "type": "array", "items": { "$ref": "#/components/schemas/HookRun" } }
            }
          }
        ]
      },
...
    }
  }
}